requirements.txt
adal==1.2.6
pytz==2021.1
msrestazure==0.6.4
pyOpenSSL==20.0.1
azure-mgmt-compute==12.0.0
azure-mgmt-resource==12.0.0
unittest-data-provider==1.0.1
